# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing orders in an online store.",
        "title": "Online Store API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/orders": {
            "post": {
                "description": "Creates a new order with items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/common.Order"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/common.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/orders/batch-create": {
            "post": {
                "description": "Creates multiple orders in batch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Batch create orders",
                "parameters": [
                    {
                        "description": "Orders data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.V1CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.V1CreateOrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/orders/query": {
            "post": {
                "description": "Query orders with filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Query orders",
                "parameters": [
                    {
                        "description": "Query filters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.V1QueryOrdersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.V1QueryOrdersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "Retrieves an order with its items by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get an order by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Order": {
            "type": "object",
            "required": [
                "customer_id",
                "delivery_address",
                "total_price_cents",
                "total_price_currency"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "integer",
                    "minimum": 0
                },
                "delivery_address": {
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.OrderItem"
                    }
                },
                "total_price_cents": {
                    "type": "integer",
                    "minimum": 0
                },
                "total_price_currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR"
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "common.OrderItem": {
            "type": "object",
            "required": [
                "price_cents",
                "price_currency",
                "product_id",
                "product_title",
                "product_url",
                "quantity"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "price_cents": {
                    "type": "integer",
                    "minimum": 0
                },
                "price_currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR"
                    ]
                },
                "product_id": {
                    "type": "integer",
                    "minimum": 0
                },
                "product_title": {
                    "type": "string",
                    "maxLength": 255
                },
                "product_url": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 0
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.V1CreateOrder": {
            "type": "object",
            "required": [
                "customer_id",
                "delivery_address",
                "order_items",
                "total_price_cents",
                "total_price_currency"
            ],
            "properties": {
                "customer_id": {
                    "type": "integer"
                },
                "delivery_address": {
                    "type": "string",
                    "maxLength": 255
                },
                "order_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.V1CreateOrderItem"
                    }
                },
                "total_price_cents": {
                    "type": "integer"
                },
                "total_price_currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR"
                    ]
                }
            }
        },
        "dto.V1CreateOrderItem": {
            "type": "object",
            "required": [
                "price_cents",
                "price_currency",
                "product_id",
                "product_title",
                "product_url",
                "quantity"
            ],
            "properties": {
                "price_cents": {
                    "type": "integer"
                },
                "price_currency": {
                    "type": "string",
                    "enum": [
                        "USD",
                        "EUR"
                    ]
                },
                "product_id": {
                    "type": "integer"
                },
                "product_title": {
                    "type": "string",
                    "maxLength": 255
                },
                "product_url": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "dto.V1CreateOrderRequest": {
            "type": "object",
            "required": [
                "orders"
            ],
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.V1CreateOrder"
                    }
                }
            }
        },
        "dto.V1CreateOrderResponse": {
            "type": "object",
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.Order"
                    }
                }
            }
        },
        "dto.V1QueryOrdersRequest": {
            "type": "object",
            "properties": {
                "customer_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "include_order_items": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "dto.V1QueryOrdersResponse": {
            "type": "object",
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.Order"
                    }
                }
            }
        }
    }
}